added catalog-heading test. It doesn't need e2e so I move on.
